+++ python/lsst/ts/ofc/ofc.py
@@ -25,9 +25,11 @@ import logging
 
 import numpy as np
 
-from . import BendModeToForce, Correction, OFCController, StateEstimator
+from . import BendModeToForce, Correction, StateEstimator
+from .controllers import BaseController, OICController, PIDController
 from .ofc_data import OFCData
-from .utils import CorrectionType
+from .utils import CorrectionType, get_filter_name
+from .utils.ofc_data_helpers import get_sensor_names
 
 
 class OFC:
@@ -57,10 +59,10 @@ class OFC:
         `set_pssn_gain()`.
     log : `logging.Logger`
         Logger class used for logging operations.
-    lv_dof : `np.ndarray`
+    lv_dof : `np.ndarray[float]`
         Last visit degrees of freedom.
-    ofc_controller : `OFCController`
-        Instance of `OFCController` class.
+    controller : `PIDController` or `OICController`
+        Controller class.
     ofc_data : `OFCData`
         OFC data container.
     pssn_data : `dict`
@@ -70,7 +72,6 @@ class OFC:
 
     Notes
     -----
-
     FWHM: Full width at half maximum.
     PSSN: Normalized point source sensitivity.
     """
@@ -81,32 +82,28 @@ class OFC:
         else:
             self.log = log.getChild(type(self).__name__)
 
-        self.pssn_data = dict(sensor_names=None, pssn=None)
-
         self.ofc_data = ofc_data
 
         self.state_estimator = StateEstimator(self.ofc_data)
 
-        self.ofc_controller = OFCController(self.ofc_data)
-
-        self.fwhm_threshold = 0.2
-
-        self.default_gain = 0.7
+        if self.ofc_data.controller["name"] == "PID":
+            self.controller: BaseController = PIDController(self.ofc_data)
+        elif self.ofc_data.controller["name"] == "OIC":
+            self.controller = OICController(self.ofc_data)
 
         # Truncation of degenerate modes after correction calculation
         self.rcond_degeneracy = 1e-3
 
         # Last visit dof
-        self.lv_dof = self.ofc_controller.dof_state.copy()
+        self.lv_dof = self.controller.dof_state.copy()
 
         self.dof_order = ("m2HexPos", "camHexPos", "M1M3Bend", "M2Bend")
 
     def calculate_corrections(
         self,
-        wfe: np.ndarray,
-        sensor_names: list[str],
+        wfe: np.ndarray[float],
+        sensor_ids: np.ndarray[int],
         filter_name: str,
-        gain: float,
         rotation_angle: float,
     ) -> list[Correction]:
         """Calculate the Hexapod, M1M3, and M2 corrections from the FWHM
@@ -114,19 +111,14 @@ class OFC:
 
         Parameters
         ----------
-        wfe : `np.ndarray`
+        wfe : `np.ndarray[float]`
             An array of arrays (e.g. 2-d array) with wavefront erros. Each
             element contains an array of wavefront errors (in um) for a
             particular detector/field.
-        sensor_names: `list`
-            List of sensor names.
+        sensor_ids: np.ndarray[int]
+            List of sensor ids.
         filter_name : `string`
-            Name of the filter used in the observations. This must be a valid
-            entry in the `ofc_data.intrinsic_zk` and `ofc_data.eff_wavelength`
-            dictionaries.
-        gain : `float`
-            User provided gain. If < 0, calculate gain based on point source
-            sensitivity normalized (PSSN).
+            Name of the filter used in the observations.
         rotation_angle : `float`
             Camera rotator angle (in degrees) during the observations.
 
@@ -142,29 +134,29 @@ class OFC:
             If size of `wfe` is different than `sensor_names`.
         """
 
-        if len(wfe) != len(sensor_names):
+        if len(wfe) != len(sensor_ids):
             RuntimeError(
                 f"Number of wavefront errors ({len(wfe)}) must be the same as "
-                f"number of sensors ({len(sensor_names)})."
+                f"number of sensors ({len(sensor_ids)})."
             )
-        # Set the gain value
-        if gain < 0.0:
-            self.set_pssn_gain()
-        else:
-            self.ofc_controller.gain = gain
+
+        # Process filter name to be in the correct format.
+        filter_name = get_filter_name(filter_name)
+        # Get sensor names from sensor ids
+        sensor_names = get_sensor_names(ofc_data=self.ofc_data, sensor_ids=sensor_ids)
 
         optical_state = self.state_estimator.dof_state(
             filter_name, wfe, sensor_names, rotation_angle
         )
 
         # Calculate the uk based on the control algorithm
-        uk = self.ofc_controller.uk_gain(filter_name, optical_state, sensor_names)
+        uk = self.controller.control_step(filter_name, optical_state, sensor_names)
 
         # Assign the value to the last visit DOF
         self.set_last_visit_dof(uk)
 
         # Aggregate the rotated uk
-        self.ofc_controller.aggregate_state(uk, self.ofc_data.dof_idx)
+        self.controller.aggregate_state(uk, self.ofc_data.dof_idx)
 
         return self.get_all_corrections()
 
@@ -201,7 +193,7 @@ class OFC:
         end_idx = start_idx + self.ofc_data.comp_dof_idx[dof_comp]["idxLength"]
         dof_idx = np.arange(start_idx, end_idx)
 
-        dof = self.ofc_controller.dof_state[dof_idx]
+        dof = self.controller.dof_state[dof_idx]
 
         if isinstance(self.ofc_data.comp_dof_idx[dof_comp]["rot_mat"], float):
             trans_dof = self.ofc_data.comp_dof_idx[dof_comp]["rot_mat"] * dof
@@ -223,36 +215,27 @@ class OFC:
     def init_lv_dof(self) -> None:
         """Initialize last visit degree of freedom."""
 
-        self.lv_dof = np.zeros_like(self.ofc_controller.dof_state0)
+        self.lv_dof = np.zeros_like(self.controller.dof_state0)
 
-    def set_fwhm_data(self, fwhm, sensor_names):
+    def set_fwhm_data(
+        self, fwhm: np.ndarray[float], sensor_ids: np.ndarray[int]
+    ) -> None:
         """Set the list of FWHMSensorData of each CCD of camera.
         Parameters
         ----------
-        fwhm : `np.ndarray`
+        fwhm : `np.ndarray[float]`
             Array of arrays (e.g. 2-d array) which contains the FWHM data.
             Each element contains an array of fwhm (in arcsec) measurements for
             a  particular sensor.
-        sensor_names : `list` of `string`
-            List of sensor names.
-        Raises
-        ------
-        RuntimeError
-            If size of `fwhm` and `sensor_names` are different.
+        sensor_ids : `np.ndarray[int]`
+            List of sensor ids.
         """
 
-        if len(fwhm) != len(sensor_names):
-            raise RuntimeError(
-                f"Size of fwhm ({len(fwhm)}) is different than sensor_names ({len(sensor_names)})."
-            )
+        # Get sensor names from sensor ids
+        sensor_names = get_sensor_names(ofc_data=self.ofc_data, sensor_ids=sensor_ids)
 
-        self.pssn_data["sensor_names"] = sensor_names.copy()
-        self.pssn_data["pssn"] = np.zeros(len(fwhm))
-
-        for s_id, fw in enumerate(fwhm):
-            self.pssn_data["pssn"][s_id] = np.average(
-                self.ofc_controller.fwhm_to_pssn(fwhm)
-            )
+        # Delegate the call to the controller's set_fwhm_data
+        self.controller.set_fwhm_data(fwhm, sensor_names)
 
     def reset(self) -> list[Correction]:
         """Reset the OFC calculation state, which is the aggregated DOF now.
@@ -273,36 +256,12 @@ class OFC:
             The figure offset for the MT M2.
         """
 
-        self.ofc_controller.reset_dof_state()
+        self.controller.reset_dof_state()
         self.init_lv_dof()
 
         return self.get_all_corrections()
 
-    def set_pssn_gain(self):
-        """Set the gain value based on the PSSN, which comes from the FWHM by
-        DM team.
-
-        Raises
-        ------
-        RuntimeError
-            If `pssn_data` is not properly set.
-        """
-
-        if self.pssn_data["pssn"] is None or self.pssn_data["sensor_names"] is None:
-            raise RuntimeError(
-                "PSSN data not set. Run `set_fwhm_data` with appropriate data."
-            )
-
-        fwhm_gq = self.ofc_controller.effective_fwhm_g4(
-            self.pssn_data["pssn"], self.pssn_data["sensor_names"]
-        )
-
-        if fwhm_gq > self.fwhm_threshold:
-            self.ofc_controller.gain = 1.0
-        else:
-            self.ofc_controller.gain = self.default_gain
-
-    def set_last_visit_dof(self, dof) -> None:
+    def set_last_visit_dof(self, dof: np.ndarray[float]) -> None:
         """Set the state (or degree of freedom, DOF) correction from the last
         visit.
 
@@ -312,7 +271,7 @@ class OFC:
             Calculated degrees of freedom.
         """
 
-        lv_dof = np.zeros_like(self.ofc_controller.dof_state0)
+        lv_dof = np.zeros_like(self.controller.dof_state0)
 
         lv_dof[self.ofc_data.dof_idx] = dof