+++ python/lsst/ts/ofc/ofc_data/ofc_data.py
@@ -23,6 +23,7 @@ __all__ = ["OFCData"]
 
 import asyncio
 import logging
+import typing
 from pathlib import Path
 
 import numpy as np
@@ -61,6 +62,8 @@ class OFCData(BaseOFCData):
         other files when the name is set.
     config_dir : `pathlib.Path`
         Path to the directory storing configuration files.
+    controller_filename : `string`
+        Controller configuration filename.
     dof_idx : `dict` of `string`
         Index of Degree of Freedom (DOF).
     field_idx : `dict` of `string`
@@ -81,8 +84,6 @@ class OFCData(BaseOFCData):
         Sensitivity matrix M.
     start_task : `asyncio.Future`
         Asyncio future that tracks whether the class is setup and ready or not.
-    xref : `string`
-        Define how the control strategy will handle the reference point.
     xref_list : `list` of `string`
         Available reference point strategies.
     y2_correction : `np.ndarray` of `float`
@@ -100,8 +101,14 @@ class OFCData(BaseOFCData):
         If input `config_dir` does not exists.
     """
 
-    def __init__(self, name=None, config_dir=None, log=None, **kwargs):
-        super().__init__(**kwargs)
+    def __init__(
+        self,
+        name: str | None = None,
+        config_dir: str | None = None,
+        log: logging.Logger | None = None,
+        **kwargs: dict[str, typing.Any],
+    ) -> None:
+        super().__init__(**kwargs)  # type: ignore
 
         # Set logger
         if log is None:
@@ -141,13 +148,21 @@ class OFCData(BaseOFCData):
         # create a new event loops and try again.
         try:
             self._configure_lock = asyncio.Lock()
-            self.start_task = asyncio.Future()
+            self.start_task: asyncio.Future = asyncio.Future()
         except RuntimeError:
             asyncio.set_event_loop(asyncio.new_event_loop())
             self._configure_lock = asyncio.Lock()
             self.start_task = asyncio.Future()
 
         self.start_task.set_result(None)
+
+        # Initialize controller configuration
+        self._controller_filename = "oic_controller.yaml"
+
+        # Zernike indices used
+        self._zn_idx = np.arange(self.znmax - self.znmin + 1, dtype=int)
+        self._zn_idx_mask = np.ones_like(self._zn_idx, dtype=bool)
+        self._zn_selected = np.arange(self.znmin, self.znmax + 1, dtype=int)
 
         # Set the name of the instrument. This reads the instrument-related
         # configuration files.
@@ -193,16 +208,8 @@ class OFCData(BaseOFCData):
         )
         self._dof_idx_mask = np.ones_like(self._dof_idx, dtype=bool)
 
-        # Zernike indices used
-        self._zn_idx = np.arange(self.znmax - self.znmin + 1, dtype=int)
-        self._zn_idx_mask = np.ones_like(self._zn_idx, dtype=bool)
-        self._zn_selected = np.arange(self.znmin, self.znmax + 1, dtype=int)
-
-        # Control strategy
-        self._xref = None
-
     @property
-    def name(self):
+    def name(self) -> str | None:
         if not self.start_task.done() or self.start_task.result() is None:
             raise RuntimeError(
                 "Class not setup or still loading instrument files."
@@ -211,7 +218,20 @@ class OFCData(BaseOFCData):
         return self.start_task.result()
 
     @name.setter
-    def name(self, value):
+    def name(self, value: str) -> None:
+        """Set the name of the instrument.
+
+        Parameters
+        ----------
+        value : `string`
+            Name of the instrument.
+
+        Raises
+        ------
+        RuntimeError
+            If instrument being configured. Cannot interrupt process. Wait for
+            `start_task` to complete before setting instrument again.
+        """
         if not self.start_task.done():
             raise RuntimeError(
                 "Instrument being configured. Cannot interrupt process. "
@@ -222,14 +242,34 @@ class OFCData(BaseOFCData):
             self._configure_instrument(value)
 
     @property
-    def xref(self):
+    def xref(self) -> str:
+        """Reference point strategy.
+
+        Returns
+        -------
+        `string`
+            Reference point strategy.
+        """
         if self._xref is None:
             return "x00"
         else:
             return self._xref
 
     @xref.setter
-    def xref(self, value):
+    def xref(self, value: str) -> None:
+        """Set the reference point strategy.
+
+        Parameters
+        ----------
+        value : `string`
+            Reference point strategy.
+
+        Raises
+        ------
+        ValueError
+            If value is not in the list of available reference
+            point strategies.
+        """
         if value in self.xref_list:
             self._xref = value
         else:
@@ -238,27 +278,41 @@ class OFCData(BaseOFCData):
             )
 
     @property
-    def xref_list(self):
+    def xref_list(self) -> set[str]:
+        """Available reference point strategies."""
         return {"x00", "x0", "0"}
 
     # Properties to access the Zernike indices
     @property
-    def zn_idx(self):
+    def zn_idx(self) -> np.ndarray[int]:
         """Zernike indices used."""
         return self._zn_idx[self.zn_idx_mask]
 
     @property
-    def zn_idx_mask(self):
+    def zn_idx_mask(self) -> np.ndarray[bool]:
         """Mask to select Zernike indices."""
         return self._zn_idx_mask
 
     @property
-    def zn_selected(self):
+    def zn_selected(self) -> np.ndarray[int]:
         """Zernike indices selected in Zernike Noll Index."""
         return self._zn_selected
 
     @zn_selected.setter
-    def zn_selected(self, value):
+    def zn_selected(self, value: np.ndarray[int]) -> None:
+        """Set the Zernike indices selected in Zernike Noll Index.
+
+        Parameters
+        ----------
+        value : `np.ndarray`
+            Zernike indices selected in Zernike Noll Index.
+
+        Raises
+        ------
+        ValueError
+            If any element in the value vector is smaller or larger than
+            `znmin` and `znmax`.
+        """
         # Check if any element in the value vector is
         # smaller or larger than zmin and zmax
         if np.any(np.array(value) < self.znmin) or np.any(np.array(value) > self.znmax):
@@ -269,19 +323,38 @@ class OFCData(BaseOFCData):
         self._zn_idx_mask = np.isin(self._zn_idx, self._zn_selected - self.znmin)
 
     @property
-    def dof_idx(self):
+    def dof_idx(self) -> np.ndarray[int]:
+        """Index of Degree of Freedom (DOF)."""
         return self._dof_idx[self.dof_idx_mask]
 
     @property
-    def dof_idx_mask(self):
+    def dof_idx_mask(self) -> np.ndarray[bool]:
+        """Mask to select Degree of Freedom (DOF)."""
         return self._dof_idx_mask
 
     @property
-    def comp_dof_idx(self):
+    def comp_dof_idx(self) -> dict:
+        """Index of Degree of Freedom (DOF) per component."""
         return self._comp_dof_idx
 
     @comp_dof_idx.setter
-    def comp_dof_idx(self, value):
+    def comp_dof_idx(self, value: dict) -> None:
+        """Set the index of Degree of Freedom (DOF) per component.
+
+        Parameters
+        ----------
+        value : `dict`
+            Index of Degree of Freedom (DOF) per component.
+
+        Raises
+        ------
+        ValueError
+            If value is not a dictionary with the expected keys.
+        RuntimeError
+            If size of input vector is different than expected.
+        RuntimeError
+            If input is not a numpy array of type bool.
+        """
         if not isinstance(value, dict):
             raise ValueError(
                 f"comp_dof_idx must be a dictionary with {self.comp_dof_idx.keys()} entries."
@@ -303,12 +376,35 @@ class OFCData(BaseOFCData):
                 raise RuntimeError("Input should be np.ndarray of type bool.")
             self._dof_idx_mask[start_idx:end_idx] = value[comp]
 
-    def load_yaml_file(self, file_path: str) -> dict:
+    @property
+    def controller_filename(self) -> str:
+        """Controller configuration filename.
+
+        Returns
+        -------
+        `string`
+            Controller configuration filename.
+        """
+        return self._controller_filename
+
+    @controller_filename.setter
+    def controller_filename(self, value: str) -> None:
+        """Set the controller configuration filename.
+
+        Parameters
+        ----------
+        value : `string`
+            Controller configuration filename.
+        """
+        self._controller_filename = value
+        self.configure_controller()
+
+    def load_yaml_file(self, file_path: Path | str) -> dict:
         """Load yaml file.
 
         Parameters
         ----------
-        file_path : `string`
+        file_path : `pathlib.Path` or `string`
             Path to the yaml file.
 
         Returns
@@ -344,7 +440,7 @@ class OFCData(BaseOFCData):
             loop = asyncio.get_running_loop()
             await loop.run_in_executor(None, self._configure_instrument, instrument)
 
-    def _configure_instrument(self, instrument):
+    def _configure_instrument(self, instrument: str) -> None:
         """Configure OFCData instrument.
 
         Parameters
@@ -376,9 +472,17 @@ class OFCData(BaseOFCData):
             If sensitivity matrix file does not exist.
         RuntimeError
             If configuration file does not exist.
-
+        RuntimeError
+            If controller configuration file does not exist.
+        ValueError
+            If controller name is missing in the controller configuration.
+        ValueError
+            If controller name is not PID or OIC.
+        ValueError
+            If required key is missing in the PID controller configuration.
+        ValueError
+            If required key is missing in the OIC controller configuration.
         """
-
         # Set the name of the instrument. Either lsst or comcam
         camera_type = instrument if (instrument != "lsstfam") else "lsst"
 
@@ -399,6 +503,11 @@ class OFCData(BaseOFCData):
                         self.bend_mode[comp][ftype]["data"] = yaml.safe_load(fp)
 
         self.log.debug(f"Configuring {instrument}")
+
+        # Sensor id to name mapping
+        self.sensor_id_to_name = self.load_yaml_file(
+            self.config_dir / "sensor_ids_to_names.yaml"
+        )
 
         # Load all data to local variables and only set them at the end if
         # everthing went fine. Otherwise you can leave the class in a broken
@@ -523,9 +632,12 @@ class OFCData(BaseOFCData):
 
         config = self.load_yaml_file(configuration_path)
 
-        self.log.debug(f"Done configuring {instrument}")
+        # Read configuration file for controller
+        # ---------------------------------------
+        self.configure_controller()
 
         # Now all data was read successfully, time to set it up.
+        # ------------------------------------------------------
         self.alpha = alpha
         self.config = config
         self.dof_state0 = dof_state0
@@ -539,8 +651,51 @@ class OFCData(BaseOFCData):
         self.sensitivity_matrix = sensitivity_matrix
         self.start_task.set_result(instrument)
 
-    async def __aenter__(self):
+        self.log.debug(f"Done configuring {instrument}")
+
+    def configure_controller(self) -> None:
+        """Refresh the controller configuration based
+        on the current controller_filename.
+        """
+        if Path(self.controller_filename).is_absolute():
+            controller_path = Path(self.controller_filename)
+        else:
+            controller_path = (
+                self.config_dir / "configurations" / self.controller_filename
+            )
+
+        self.controller = self.load_yaml_file(controller_path)
+
+        if "name" not in self.controller:
+            raise ValueError(
+                "Required key 'name' is missing in the controller configuration."
+            )
+
+        if self.controller["name"] not in ["PID", "OIC"]:
+            raise ValueError("Controller 'name' must be either 'PID' or 'OIC'.")
+
+        # check if the zn_selected key is present in
+        # the controller configuration
+        if "zn_selected" in self.controller:
+            self.zn_selected = np.array(self.controller["zn_selected"])
+
+        if self.controller["name"] == "PID":
+            for key in ["kp", "ki", "kd", "setpoint"]:
+                if key not in self.controller:
+                    raise ValueError(
+                        f"Required key '{key}' is missing in the PID controller configuration."
+                    )
+        elif self.controller["name"] == "OIC":
+            for key in ["kp", "ki", "kd", "setpoint", "xref"]:
+                if key not in self.controller:
+                    raise ValueError(
+                        f"Required key '{key}' is missing in the OIC controller configuration."
+                    )
+
+            self.xref = self.controller["xref"]
+
+    async def __aenter__(self) -> "OFCData":
         return self
 
-    async def __aexit__(self, type, value, traceback):
-        await self.close()
+    async def __aexit__(self, type: object, value: object, traceback: object) -> None:
+        pass