+++ python/lsst/ts/ofc/utils/utils.py
@@ -19,34 +19,55 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-__all__ = ["get_pkg_root", "get_config_dir", "rot_1d_array"]
+__all__ = ["get_pkg_root", "get_config_dir", "get_filter_name", "rot_1d_array"]
 
 import pathlib
 
+import numpy as np
 
-def get_pkg_root():
+
+def get_pkg_root() -> pathlib.Path:
     """Return the root directory of this package.
 
     Returns
     -------
-    `pathlib.PosixPath`
+    `pathlib.Path`
         Path to the package root.
     """
     return pathlib.Path(__file__).resolve().parents[5]
 
 
-def get_config_dir():
+def get_config_dir() -> pathlib.Path:
     """Return the path to the ``policy`` dir within this package.
 
     Returns
     -------
-    `pathlib.PosixPath`
+    `pathlib.Path`
         Path to the package config directory.
     """
     return get_pkg_root() / "policy"
 
 
-def rot_1d_array(array, rot_mat):
+def get_filter_name(filter_name: str) -> str:
+    """Return the filter name in the format used by the OFC.
+
+    Parameters
+    ----------
+    filter_name : `str`
+        Filter name.
+
+    Returns
+    -------
+    `str`
+        Filter name in the format used by the OFC.
+    """
+    if "_" in filter_name:
+        return filter_name.split("_")[0].upper()
+    else:
+        return filter_name.upper()
+
+
+def rot_1d_array(array: np.ndarray[float], rot_mat: np.ndarray[float]) -> np.ndarray:
     """Rotate 1D array from one basis to another.
 
     Parameters