+++ doc/versionHistory.rst
@@ -6,6 +6,20 @@
 Version History
 ##################
 
+.. _lsst.ts.ofc-3.2.0:
+
+v3.2.0
+======
+
+* Add PIDController as alternative control strategy.
+* Move current control strategy to OICController (optimal integral controller).
+* Add PID option to OICController output.
+* Add mypy type checking to ts_ofc.
+* Allow filter names to be passed as 'r_01', 'r' or 'R'.
+* Change `ofc.calculate_corrections()` to expect sensor_ids instead of sensor_names.
+* Add comcam to policy/sensor_ids_to_names.yaml.
+* Add zn_selected as a configuration too policy files.
+
 .. _lsst.ts.ofc-3.1.3:
 
 v3.1.3